Move name change signals over to TreeHandler
